+++ client/src/app/(protected)/groups/[id]/Content/Sessions/CardSession.tsx
@@ -1,43 +1,11 @@
 import GroupSizeIcon from '@/assets/Icons/GroupSizeIcon';
 import LocationIcon from '@/assets/Icons/LocationIcon';
+import { dateFormat } from '@/utils/Formatter';
 
 type CardSessionProps = {
   session: any; // that type will specific when conect this to back
   isHistory?: boolean;
 };
-
-function dateFormat(startTime: string, endTime: string, isHistory: boolean) {
-  const startDate = startTime.split('T')[0];
-  const endDate = endTime.split('T')[0];
-  const endHour = endTime.split('T')[1].split(':');
-  const startHour = startTime.split('T')[1].split(':');
-
-  const day = startDate.split('-')[2];
-  const month = Number(startDate.split('-')[1]);
-  const year = startDate.split('-')[0];
-
-  const showEndHour =
-    isHistory && startDate === endDate && endHour[0] !== startHour[0]
-      ? `/ ${endHour[0]}:${endHour[1]}`
-      : '';
-
-  const months: { [key: number]: string } = {
-    1: 'Enero',
-    2: 'Febrero',
-    3: 'Marzo',
-    4: 'Abril',
-    5: 'Mayo',
-    6: 'Junio',
-    7: 'Julio',
-    8: 'Agosto',
-    9: 'Septiembre',
-    10: 'Octubre',
-    11: 'Noviembre',
-    12: 'Diciembre',
-  };
-
-  return `${day} ${months[month]} ${year} ${startHour[0]}:${startHour[1]} ${showEndHour}`;
-}
 
 export default function CardSession({
   session,
